ui/requirements.txt
streamlit==1.47.1
websockets==15.0.1
websocket-client==1.8.0